pyproject.toml
[project]
name = "simplifai-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.0",
    "google-genai>=1.11.0",
    "pillow>=11.2.1",
    "pymupdf>=1.25.5",
    "pytesseract>=0.3.13",
    "python-docx>=1.0.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "pywin32>=306; platform_system == 'Windows'",
    "regex>=2023.0.0",
    "scikit-image>=0.21.0",
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "langdetect>=1.0.9",
]
